(refactor) test configuration 분리
